move data models to a separate directory
