+++ Info-Orbs/include/model/stockDataModel.h
