+++ Info-Orbs/include/model/weatherDataModel.h
