+++ Info-Orbs/include/widgets/weatherWidget.h
@@ -10,7 +10,7 @@
 #include <widget.h>
 
 #include "icons.h"
-#include "weatherDataModel.h"
+#include "model/weatherDataModel.h"
 
 class WeatherWidget : public Widget {
    public:
